% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_test.R
\name{adj_chisq}
\alias{adj_chisq}
\title{Adjusted Chi-Square Test Statistic}
\usage{
adj_chisq(test_stat, config)
}
\arguments{
\item{test_stat}{Numeric value representing the chi-square test statistic.}

\item{config}{A list containing the configuration parameters:
\describe{
\item{q}{Parameter of interest on the boundary.}
\item{r}{Parameter of interest not on the boundary.}
\item{s}{Nuisance parameter on the boundary (in this function, it is always 0).}
\item{t}{Nuisance parameter not on the boundary.}
}}
}
\value{
A numeric value representing the calculated p-value based on the
specified configuration.
}
\description{
This function calculates the p-value for a chi-square test statistic based on the
configuration parameters passed in \code{config}. It supports different cases for
parameters on and off the boundary.
}
\details{
The function evaluates the test statistic based on three cases:
\describe{
\item{Case I}{If \code{q == 0} and \code{s == 0}, it calculates the p-value
using \code{r} degrees of freedom.}
\item{Case II}{If \code{q == 1}, \code{r == 0}, and \code{s == 0}, it
calculates a weighted p-value combining \code{df = 1} and \code{test_stat == 0}.}
\item{Case III}{If \code{q == 1} and \code{s == 0}, it calculates a weighted
p-value using degrees of freedom \code{r - 1} and \code{r}.}
}
}
\references{
Self, S.G., & Liang, K. (1987). \emph{On Asymptotic properties of maximum
likelihood estimators
and likelihood ratio tests under non-standard conditions}.
Journal of American Statistical Association, 82(398), 605-610-145.
\url{https://www.jstor.org/stable/2289471}
}
