% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{adjustedc_mixture_ci}
\alias{adjustedc_mixture_ci}
\title{Confidence Interval Computation for VPC using Mixture Normal Quantiles}
\usage{
adjustedc_mixture_ci(vpcObj, alpha = 0.05, n = 1000)
}
\arguments{
\item{vpcObj}{An object containing the fitted model and additional VPC-related information.}

\item{alpha}{Numeric. The significance level for the confidence interval (default is 0.05).}

\item{n}{Integer. Number of samples to draw for quantile estimation (default is 1000).}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates the confidence interval for the Variance Partition Coefficient (VPC)
by computing quantiles from a mixture of normal distributions. The function extracts
necessary parameters from a fitted model object and applies the \code{qmixtnorm} function.
}
