% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVpc.R
\name{batchGLMMFit}
\alias{batchGLMMFit}
\title{Fit Multiple Generalized Linear Mixed Models (GLMMs)}
\usage{
batchGLMMFit(
  formula,
  dataMat,
  X,
  group,
  family,
  cov_values,
  nulltest = NULL,
  parallel = TRUE,
  num_cores = NULL
)
}
\arguments{
\item{formula}{A formula specifying the fixed and random effects in the model.}

\item{dataMat}{A matrix where each row represents a separate dataset to fit the model to.}

\item{X}{A numeric vector or matrix of covariates for the fixed effects.}

\item{group}{A factor or numeric vector indicating the grouping structure for random effects.}

\item{family}{A character string specifying the family of distributions to use (e.g., "nb" for negative binomial).}

\item{cov_values}{A vector of covariate values at which to evaluate the Variance Partition Coefficient (VPC).}

\item{nulltest}{An optional argument for specifying a null hypothesis test (default is NULL).}

\item{parallel}{Logical, whether to run the fitting in parallel. Default is TRUE.}

\item{num_cores}{Integer, the number of cores to use for parallel processing. Defaults to the number of available cores minus one.}
}
\value{
A list containing the following components:
\item{GlmmObj}{A list of GLMM objects, one for each row of the input data matrix.}
\item{num_feat}{The number of datasets (rows of the input matrix) that were fit.}
\item{num_covariates}{The number of covariate values used for VPC calculations.}
}
\description{
This function fits multiple Generalized Linear Mixed Models (GLMMs) to the rows of a data matrix.
Each row of the matrix is treated as a separate dataset, and the function returns a list of fitted models.
}
\examples{
ns <- c(10, 20)
X <- rgen01(ns)
beta <- c(0.5, 2)
Sigma <- matrix(c(1, 0.5, 0.5, 1), nrow=2)
family <- "nb"
cov_values <- c(0.5, 1.5)
formula <- y ~ x + (1 | group)
datafrmMat <- batchGLMMData(5, X, beta, Sigma, ns, family = "nb", theta = 1.5)
ys <- datafrmMat[-1,]; X <- datafrmMat[1,]; group <- colnames(datafrmMat)
fit <- batchGLMMFit(formula, ys, X, group, family, cov_values)
}
