% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{boostrap_vpc_ci}
\alias{boostrap_vpc_ci}
\title{Bootstrap Confidence Interval for VPC}
\usage{
boostrap_vpc_ci(vpcObj, iter = 100, num_cores = 4, alpha = 0.05)
}
\arguments{
\item{vpcObj}{An object containing model information, including coefficients, family, and grouping variable.}

\item{iter}{Integer. Number of iterations for the bootstrap simulation.}

\item{num_cores}{Integer. Number of cores to use for parallel processing during model fitting.}

\item{alpha}{Numeric. Significance level for the confidence interval (default is 0.05).}
}
\value{
A numeric vector of length 2 containing the lower and upper quantiles for the confidence interval.
}
\description{
This function calculates a bootstrap confidence interval for the Variance Partition Coefficient (VPC) by
simulating new data, fitting the model, and computing the VPC values from the bootstrap fits.
}
