% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{classical_vpc_ci}
\alias{classical_vpc_ci}
\title{Confidence Interval Calculation for VPC Using Standard Error and Critical Value}
\usage{
classical_vpc_ci(vpcObj, vpc.value, alpha = 0.05)
}
\arguments{
\item{vpcObj}{An object containing the VPC model with variance-covariance information.}

\item{vpc.value}{Numeric. The VPC value for which the confidence interval is to be computed.}

\item{alpha}{Numeric. The significance level for the confidence interval (default is 0.05).}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates the confidence interval for the Variance Partition Coefficient (VPC)
using the standard error and critical value approach. It assumes a normal approximation for the VPC.
}
