% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVpc.R
\name{coef.Glmmfit}
\alias{coef.Glmmfit}
\title{Extract Coefficients from Glmmfit Object}
\usage{
\method{coef}{Glmmfit}(object, ...)
}
\arguments{
\item{object}{A \code{Glmmfit} object that contains multiple GLMM fit results.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data frame where each row represents the coefficients from one of the fitted GLMM models.
}
\description{
This function extracts coefficients from a batch of GLMM fit objects stored within a \code{Glmmfit} object.
}
\examples{
ns <- c(10, 20)
X <- rgen01(ns)
beta <- c(0.5, 2)
Sigma <- matrix(c(1, 0.5, 0.5, 1), nrow=2)
family <- "nb"
cov_values <- c(0.5, 1.5)
formula <- y ~ x + (1 | group)
datafrmMat <- batchGLMMData(5, X, beta, Sigma, ns, family = "nb", theta = 1.5)
ys <- datafrmMat[-1,]; X <- datafrmMat[1,]; group <- colnames(datafrmMat)
fit <- batchGLMMFit(formula, ys, X, group, family, cov_values)
coefs <- coef(fit)
}
