% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_glmm_utils.R
\name{compute_mu_sig}
\alias{compute_mu_sig}
\title{Compute mu and sigma for the given design matrices and parameters}
\usage{
compute_mu_sig(beta, Sigma, x = NULL)
}
\arguments{
\item{beta}{A vector of regression coefficients.}

\item{Sigma}{A covariance matrix or vector representing variance.}

\item{x}{vector for fixed effects. If \code{x} is NULL, X should only
include the intercept.}
}
\value{
A list containing the computed \code{mu} and \code{sigma}.
}
\description{
This helper function computes the linear predictor (mu) and random effect
variance (sigma) based on the design matrices for fixed and random effects.
If \code{X} is a vector, it prepends a \code{1} to represent the intercept and ensures
\code{X} has the same length as \code{beta}. If \code{X} is NULL, only the intercept is used.
The function handles cases where \code{Sigma} is either a vector (scalar variance)
or a matrix (covariance matrix).
}
