% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_random_effects.R
\name{computeFixedEffects}
\alias{computeFixedEffects}
\title{Compute Fixed Effects Contribution to the Linear Predictor}
\usage{
computeFixedEffects(beta, X = NULL, add_intercept = TRUE)
}
\arguments{
\item{beta}{A vector of fixed effect coefficients. If \code{X} is provided,
\code{beta} should have the same number of elements as columns in \code{X}. If \code{X}
is \code{NULL}, \code{beta} should be a scalar (for an intercept-only model).}

\item{X}{A design matrix (without or with an intercept). If \code{NULL},
the model is assumed to be intercept-only.}

\item{add_intercept}{Logical to include intercept or not.}
}
\value{
A vector representing the fixed effects contribution to the
linear predictor. If \code{X} is \code{NULL} and no valid intercept is provided,
it returns \code{0}.
}
\description{
This function computes the fixed effects part of the
linear predictor in a regression model. If the design matrix \code{X} is provided,
it calculates \code{X \%*\% beta}. If \code{X} is \code{NULL} but \code{beta} is a scalar
(representing an intercept-only model), the function returns a vector
of length \code{n} with the intercept value repeated. The function checks whether
an intercept is already included in \code{X}, and if not, it adds a column of
1's for the intercept.
}
\examples{
# Example with X and beta
X <- matrix(c(1.5, 2.0, 3.1, 4.0, 5.5, 6.1, 7.3), nrow = 7, ncol = 1)
beta <- c(0.5, 1.2)  # Intercept and slope
computeFixedEffects(beta, X)

# Intercept-only example
computeFixedEffects(0.5)
computeFixedEffects(beta,X)
}
