% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_glmm_utils.R
\name{computeMuGLMM}
\alias{computeMuGLMM}
\title{Compute the linear predictor and expected response for a GLM/GLMM}
\usage{
computeMuGLMM(
  beta,
  Sigma = NULL,
  ns = NULL,
  X = NULL,
  link = "identity",
  warn = TRUE
)
}
\arguments{
\item{beta}{A numeric vector representing the fixed effect coefficients.}

\item{Sigma}{A scalar or symmetric positive-definite matrix representing the
covariance matrix of the distribution.}

\item{ns}{A vector of group sizes, indicating the number of
observations per group.}

\item{X}{A numeric matrix representing the design matrix for the fixed effects.}

\item{link}{A character string specifying the link function to use. Supported
options are \code{"identity"}, \code{"log"}, \code{"logit"}, and \code{"inverse"}.}

\item{warn}{Logical, controls warning when random effect is not computed.}
}
\value{
A numeric vector representing the expected response \eqn{\mu} after
applying the inverse link function.
}
\description{
This function computes the linear predictor (\eqn{\eta = X\beta + ZU}) for a
generalized linear model (GLM) or generalized linear mixed model (GLMM), depending
on whether the random effects (\eqn{ZU}) are provided. If random effects are not
provided, the function defaults to a standard GLM, i.e., \eqn{\eta = X\beta}.
It also applies the inverse link function to obtain the expected response \eqn{\mu}.
}
\examples{
# Example usage for a GLM with identity link (linear model)
X <- matrix(1:10, nrow=10)
beta <- c(0.5, 1.5)  # Fixed effect coefficients
ns <- c(5,3,2)
Sigma <- matrix(c(2,1,1,2),nrow=2)
mu <- computeMuGLMM(beta,Sigma, ns, X, link = "identity")
print(mu)

}
