% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_random_effects.R
\name{computeRandomEffects}
\alias{computeRandomEffects}
\title{Compute Random Effects Contribution to the Linear Predictor}
\usage{
computeRandomEffects(Z, U)
}
\arguments{
\item{Z}{A list of random effects design matrices. Each matrix corresponds
to a random effect (e.g., intercept, slope).}

\item{U}{A matrix of random effect coefficients. Each column corresponds to
the coefficients for one random effect.}
}
\value{
A vector representing the random effects contribution to the linear
predictor.
}
\description{
This function computes the random effects part of the linear predictor in a
mixed model.
It multiplies each design matrix in the list \code{Z} by the corresponding column
in \code{U} and sums the contributions.
}
\examples{
# Example with Z and U
ns <- c(2,3,2)
X <- rnorm(sum(ns))
Z <- generateRandomDesignMatrices(X = X, ns)
U <- generateRandomEffectsCoefficients(length(ns), matrix(c(2,1,1,2),2))
computeRandomEffects(Z, U)
X1 <- matrix(rnorm(sum(ns)*2),ncol=2)
Z1 <- generateRandomDesignMatrices(X = X1, ns)
}
