% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{confint.VpcObj}
\alias{confint.VpcObj}
\title{Confidence Intervals for VPC Estimates}
\usage{
\method{confint}{VpcObj}(
  VpcObj,
  alpha = 0.05,
  type = c("classical", "bootstrap", "adjusted.s", "adjusted.c"),
  order = 1,
  iter = 100,
  num_cores = 4,
  verbose = FALSE,
  prob.type = "self"
)
}
\arguments{
\item{VpcObj}{An object of class \code{vpcObj} containing the model fit and VPC estimates.}

\item{alpha}{Significance level for the confidence intervals. Default is 0.05 (for 95\% CI).}

\item{type}{Character. Specifies the type of confidence interval to compute. Options include:
\itemize{
\item "classical": Based on the standard normal distribution.
\item "bootstrap": Using a parametric bootstrap method.
\item "adjusted.s": Using the Delta Method, adjusted for boundary conditions (Self & Liang)
\item "adjusted.c": Using the Delta Method, adjusted for boundary conditions (Chant)
}}

\item{order}{An integer indicating if the first order or second order taylo approximation should be used for the delta method}

\item{iter}{Integer. The number of bootstrap iterations to perform (for bootstrap type). Default is 100.}

\item{num_cores}{Integer. The number of cores to use for parallel computation in the bootstrap method. Default is 1.}

\item{verbose}{Logical. If \code{TRUE}, provides additional information about model convergence and the Hessian matrix's positive definiteness.}

\item{prob.type}{Method use in calculating the mixture weights for adjusted method}
}
\value{
A vector with three elements: Lower bound, the VPC estimate, and the upper bound of the confidence interval.
Additional information is included if \code{verbose = TRUE}. Returns \code{NA} if the model did not converge or if the Hessian is not positive definite.
}
\description{
This function computes confidence intervals for Variance Partition Coefficient (VPC) estimates
from an object of class \code{vpcObj}. The function supports both classical and bootstrap confidence interval types.
}
