% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmData.R
\name{createGLMMData}
\alias{createGLMMData}
\title{Generate Data from a Generalized Linear Mixed Model (GLMM)}
\usage{
createGLMMData(X, beta, Sigma, ns, family, ...)
}
\arguments{
\item{X}{A numeric vector of fixed effect covariates.}

\item{beta}{A numeric vector of fixed effect coefficients. The length of
\code{beta} should match the length of \code{X}.}

\item{Sigma}{A numeric matrix representing the covariance matrix for the
random effects.}

\item{ns}{A numeric vector specifying the number of observations per group.}

\item{family}{A character string specifying the family of distributions to
use. Supported values are \code{"nb"} for negative binomial and \code{"tw"} for
Tweedie distribution.}

\item{...}{Additional arguments passed to the family-specific random number
generation functions. For \code{"nb"} (negative binomial), use \code{theta}. For \code{"tw"}
(Tweedie), use \code{phi} and \code{power}.}
}
\value{
A numeric vector of generated data based on the specified GLMM.
}
\description{
This function generates simulated data from a Generalized Linear Mixed Model
(GLMM) using the specified fixed effects, random effects, and family of
distributions. It supports negative binomial and Tweedie distributions.
}
\examples{
ns <- c(10, 20)
X <- rgen01(ns)
beta <- c(0.5, 2)
Sigma <- matrix(c(1, 0.5, 0.5, 1), nrow=2)
data_nb <- createGLMMData(X, beta, Sigma, ns, family = "nb", theta = 1.5)
data_tw <- createGLMMData(X, beta, Sigma, ns, family = "tw", phi = 1.2, power = 1.5)
}
