% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eta}
\alias{eta}
\title{Linear Predictor (Eta)}
\usage{
eta(X, beta, Z, u)
}
\arguments{
\item{X}{A numeric matrix of fixed effect covariates. Each row represents an
observation, and each column represents a covariate.}

\item{beta}{A numeric vector of fixed effect coefficients. The length of
\code{beta} should match the number of columns in \code{X}.}

\item{Z}{A numeric matrix of random effect covariates. Each row represents
an observation, and each column represents a random effect covariate.}

\item{u}{A numeric vector of random effect coefficients. The length of \code{u}
should match the number of columns in \code{Z}.}
}
\value{
A numeric vector representing the linear predictor, with each element
corresponding to an observation.
}
\description{
This function calculates the linear predictor (often denoted as  \eqn{\eta}) in
a generalized linear model (GLM) or linear mixed model (LMM). The linear
predictor is computed as a combination of fixed effects  \eqn{\beta} and
random effects  \eqn{Z}.
}
\examples{
X <- matrix(c(1, 1, 1, 0, 1, 2), nrow=3, ncol=2)
beta <- c(0.5, 2)
Z <- matrix(c(1, 0, 1, 1, 1, 0), nrow=3, ncol=2)
u <- c(1.5, -0.5)
eta(X, beta, Z, u)
}
