% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_fitting.R
\name{extractParametersByFamily}
\alias{extractParametersByFamily}
\title{Extract model parameters by family}
\usage{
extractParametersByFamily(family, modObj)
}
\arguments{
\item{family}{A character string representing the family (e.g., "nb" for Negative Binomial).}

\item{modObj}{The fitted GLMM model object.}
}
\value{
A list of extracted model parameters, for random intercept we return
the variance and not standard deviation.
}
\description{
This function extracts the relevant model parameters (e.g., beta, Sigma, phi, power)
based on the specified family from a fitted GLMM model object.
}
