% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_random_effects.R
\name{generateAndComputeRandomEffects}
\alias{generateAndComputeRandomEffects}
\title{Generate and Compute Random Effects}
\usage{
generateAndComputeRandomEffects(ns, Sigma, X = NULL)
}
\arguments{
\item{ns}{A vector of integers representing the sample sizes in each cluster.}

\item{Sigma}{A variance-covariance matrix for the random effect coefficients.}

\item{X}{A covariate vector or matrix. If it is a vector, it will be used
as the covariate for generating random design matrices. If it is a matrix,
each column represents a different covariate.}
}
\value{
A vector representing the random effects contribution to the linear
predictor, computed by multiplying each random effect design matrix (\code{Z})
with the corresponding column in the random effect coefficient matrix (\code{U}).
}
\description{
This function generates the random effects design matrices (\code{Z}) and random
effect coefficients (\code{U}) based on the input covariate matrix (\code{X}),
sample sizes in each cluster (\code{ns}), and the variance-covariance matrix
of the random effects (\code{Sigma}). It then computes the random effects contribution
to the linear predictor using the \code{computeRandomEffects} function.
}
\examples{
# Example usage:
X <- rnorm(7)  # Covariate vector
ns <- c(2, 3, 2)  # Sample sizes in each cluster
Sigma <- diag(2)  # Variance-covariance matrix
result <- generateAndComputeRandomEffects(ns, Sigma,X)
print(result)

}
