% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_random_effects.R
\name{generateRandomEffectsCoefficients}
\alias{generateRandomEffectsCoefficients}
\title{Generate Random effects from Multivariate Normal Samples}
\usage{
generateRandomEffectsCoefficients(n, Sigma)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{Sigma}{A scalar or symmetric positive-definite matrix representing the
covariance matrix of the distribution.}
}
\value{
A matrix of dimension \code{n x p}, where \code{p} is the number of
variables (the number of columns in \code{Sigma}). Each row is a
sample from the multivariate normal distribution.
}
\description{
This function generates samples from a multivariate normal distribution with
mean 0 and a specified covariance matrix.
}
\examples{
# Define a covariance matrix
Sigma <- matrix(c(1, 0.5, 0.5, 1), ncol=2)
# Generate 5 samples from a bivariate normal distribution with 0 vector mean
generateRandomEffectsCoefficients(5, Sigma)
}
