% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cluster}
\alias{get_cluster}
\title{Extract Cluster Information}
\usage{
get_cluster(object)
}
\arguments{
\item{object}{A glmmDataMatrix or batchglmmDataMatrix object.}
}
\value{
A vector of cluster names.
}
\description{
This function extracts the cluster information from the column names of a glmmDataMatrix or batchglmmDataMatrix object.
}
