% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_fitting.R
\name{get_glmmTMBfamily}
\alias{get_glmmTMBfamily}
\title{Convert family name to glmmTMB family object}
\usage{
get_glmmTMBfamily(family)
}
\arguments{
\item{family}{A string representing the family (e.g., "nb", "poisson", "gaussian").}
}
\value{
A glmmTMB family object.
}
\description{
This function converts a string representing a GLMM family (e.g., "nb" for Negative Binomial)
into a corresponding glmmTMB family object.
}
