% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_y}
\alias{get_y}
\title{Extract Response (y)}
\usage{
get_y(object)
}
\arguments{
\item{object}{A glmmDataMatrix or batchglmmDataMatrix object.}
}
\value{
A vector of response variable values.
}
\description{
This function extracts the response variable from a glmmDataMatrix or batchglmmDataMatrix object.
}
