% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmDataMatrixNB}
\alias{glmmDataMatrixNB}
\title{Batch Data for Negative Binomial GLMM}
\format{
A matrix representing the generated batch data:
\describe{
\item{X}{Covariate matrix.}
\item{Feature}{Generated response variable based on the model.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
glmmDataMatrixNB
}
\description{
A dataset generated for a generalized linear mixed model using Negative Binomial parameters.
}
\keyword{datasets}
