% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glmmTMBfamily}
\alias{glmmTMBfamily}
\title{GLMMTMB Familiies}
\usage{
glmmTMBfamily(family)
}
\arguments{
\item{family}{A string indicating the family to be used. Available options are:
"nb" (negative binomial), "tw" (Tweedie), "ga" (Gaussian).}
}
\value{
The corresponding family function for \code{glmmTMB}.
}
\description{
This function returns the appropriate family function for \code{glmmTMB} based on the input string.
}
\examples{
glmmTMBfamily("nb")
glmmTMBfamily("tw")
glmmTMBfamily("ga")
}
