% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_fitting.R
\name{logLik.glmmfit}
\alias{logLik.glmmfit}
\title{Log-Likelihood for GLMM Fit}
\usage{
logLik.glmmfit(object, ...)
}
\arguments{
\item{object}{A fitted GLMM object of class \code{glmmfit}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The log-likelihood value of the fitted model.
}
\description{
This function extracts the log-likelihood of a fitted generalized linear mixed model (GLMM) object.
}
\examples{
\dontrun{
  fit <- glmmfit(data, family = "binomial")
  log_likelihood <- logLik(fit)
}

}
\seealso{
\code{\link[stats]{logLik}} for the generic log-likelihood method.
}
