% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logNormMoment}
\alias{logNormMoment}
\title{Compute the k-th Moment of a Log-Normal Distribution}
\usage{
logNormMoment(mu, Sigma, k)
}
\arguments{
\item{mu}{Numeric value representing the mean of the underlying normal distribution.}

\item{Sigma}{Numeric value representing the variance (sigma^2) of the underlying normal distribution.}

\item{k}{Numeric value representing the moment to compute.}
}
\value{
Returns the k-th moment of the log-normal distribution.
}
\description{
This function computes the k-th moment of a log-normal distribution given the mean \code{mu},
variance \code{Sigma} (or sigma^2), and a constant \code{k}.
}
\examples{
logNormMoment(0, 1, 1) # First moment (mean) of a log-normal distribution
logNormMoment(0.5, 2, 2) # Second moment of a log-normal distribution
}
