% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDataMatrix}
\alias{makeDataMatrix}
\title{Create a Data Matrix from Covariates, Response Values, and Clusters}
\usage{
makeDataMatrix(X, ys, cluster)
}
\arguments{
\item{X}{A matrix of covariates. If the number of columns does not match
the number of
response values, it will be transposed if the number of rows matches.}

\item{ys}{A matrix or vector of response values. If a vector, it will be
converted to a matrix.}

\item{cluster}{A character vector containing cluster names to be assigned
as column names
for the resulting data matrix.}
}
\value{
A matrix combining covariates, response values, and cluster identifiers.
The resulting matrix has the response values as the last rows and
columns named by the clusters.
}
\description{
This function combines a matrix of covariates (X), a matrix of response
values (ys),
and cluster identifiers into a single data matrix. It ensures that dimensions
are compatible
and assigns the appropriate column names.
}
