% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_glmm_utils.R
\name{null_or_default}
\alias{null_or_default}
\alias{\%||\%}
\title{Custom operator to handle NULL with a default}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The value to be checked.}

\item{b}{The default value to return if \code{a} is NULL.}
}
\value{
Returns \code{a} if it is not NULL, otherwise returns \code{b}.
}
\description{
This custom operator returns the left-hand side value unless it is NULL, in
which case it returns the right-hand side value.
}
\examples{
x <- NULL
y <- 10
x \%||\% y  # returns 10

x <- 5
x \%||\% y  # returns 5

}
