% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_test.R
\name{obtain_config}
\alias{obtain_config}
\title{Obtain Configuration Parameters for Boundary and Non-Boundary Models}
\usage{
obtain_config(null, alt)
}
\arguments{
\item{null}{A model object representing the null hypothesis. This object
should contain
\code{modObj} and \code{Sigma}, where \code{modObj} is the model object and \code{Sigma} is a
matrix indicating
the parameters on the boundary.}

\item{alt}{A model object representing the alternative hypothesis. This
object should contain
\code{modObj} and \code{Sigma}, where \code{modObj} is the model object and \code{Sigma} is a
matrix indicating
the parameters on the boundary.}
}
\value{
A list with the following components:
\describe{
\item{q}{Number of parameters of interest that are on the boundary in the
alternative model but not in the null model.}
\item{r}{Number of parameters of interest that are not on the boundary in
either the null or alternative model.}
\item{s}{Number of nuisance parameters on the boundary (always 0 in these
cases).}
\item{t}{Number of nuisance parameters not on the boundary.}
}
}
\description{
This function calculates the parameters \code{q}, \code{r}, \code{s}, and \code{t} based on the
difference
between a null and alternative model, particularly in cases where certain
parameters
are on the boundary or not.
}
\references{
Self, S.G., & Liang, K. (1987). \emph{On Asymptotic properties of maximum
likelihood estimators
and likelihood ratio tests under non-standard conditions}.
Journal of American Statistical Association, 82(398), 605-610-145.
\url{https://www.jstor.org/stable/2289471}
}
