% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmData.R
\name{parallelbatchGLMMData}
\alias{parallelbatchGLMMData}
\title{Generate GLMM Data from a Parameter Matrix}
\usage{
parallelbatchGLMMData(
  paramMat,
  X,
  ns,
  family,
  iter = 1,
  parallel = TRUE,
  num_cores = NULL
)
}
\arguments{
\item{paramMat}{A matrix where each row contains parameters for generating a GLMM dataset.}

\item{X}{A numeric vector of fixed effect covariates.}

\item{ns}{A numeric vector specifying the number of observations per group.}

\item{family}{A character string specifying the family of distributions to use.
Supported values are \code{"nb"} for negative binomial and \code{"tw"} for Tweedie distribution.}

\item{iter}{An integer specifying the number of datasets (features) to generate per parameter row.}

\item{parallel}{A logical value indicating whether to use parallel processing. Default is \code{TRUE}.}

\item{num_cores}{integer indicating number of cores to use fo fitting}
}
\value{
A numeric matrix where each row represents a generated dataset, and each column corresponds to a group specified in \code{ns}.
}
\description{
This function generates datasets from a Generalized Linear Mixed Model (GLMM)
using a matrix of parameters. The function supports parallel processing
to speed up the generation process, especially with large parameter matrices.
}
\examples{
ns <- c(10, 20)
X <- rgen01(ns)
family <- "nb"
b0 <- 3
b1 <- 7
phi <- 0.8
sig11 <- 2
sig12 <- 1
sig22 <- 2
paramMat <- expand.grid(b0=b0,b1=b1,phi=phi,sig11=sig11,sig12=sig12,sig22=sig22)
parallelbatchGLMMData(paramMat, X, ns, family, iter=3, parallel=TRUE)
}
