% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paramNBmat}
\alias{paramNBmat}
\title{Parameter Matrix for Negative Binomial}
\format{
A parameter matrix for Negative Binomial for generating glmmDataMatrix:
\describe{
\item{b0}{Values for parameter b0.}
\item{b1}{Values for parameter b1.}
\item{sig11}{Variance for the random intercept.}
\item{sig12}{Covariance between the random intercept and random slope.}
\item{sig22}{Variance for the random slope.}
\item{theta}{Dispersion parameter for the Negative Binomial model.}
}
}
\source{
Generated for use in GLMM fitting.
}
\usage{
paramNBmat
}
\description{
A dataset containing combinations of parameters for a generalized linear mixed model.
}
\keyword{datasets}
