% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{qmixtnorm}
\alias{qmixtnorm}
\title{Quantile Computation for a Mixture of Multivariate & Truncated Normals}
\usage{
qmixtnorm(mean, Sigma, pis, grad, alpha = 0.05, n = 100, truncated = TRUE)
}
\arguments{
\item{mean}{A numeric vector representing the mean of each component in the multivariate normal distribution.}

\item{Sigma}{A covariance matrix associated with the multivariate normal distribution.}

\item{pis}{A numeric vector of mixing proportions for the components of the mixture model. The elements of \code{pis} should sum to 1.}

\item{grad}{A numeric vector representing the gradient, used to transform the sampled mixture data.}

\item{alpha}{A numeric value specifying the significance level for the confidence interval (default is 0.05, which corresponds to a 95\% confidence interval).}

\item{n}{An integer specifying the number of samples to draw from the mixture distribution (default is 100).}

\item{truncated}{Logical. Whether truncation should be used.}
}
\value{
A numeric vector of length 2 containing the lower and upper quantiles for the confidence interval, calculated as \code{c(alpha/2, 1 - alpha/2)}.
}
\description{
This function calculates the quantiles for a mixture of multivariate normal
distributions
based on the provided mean vector, covariance matrix, mixture proportions,
and gradient vector.
The quantiles are returned for the confidence interval specified by \code{alpha}.
}
