% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgen01}
\alias{rgen01}
\title{Generate Random Binary Vectors with Equal Probability}
\usage{
rgen01(ns)
}
\arguments{
\item{ns}{A numeric vector where each element specifies the length of the
binary vector to be generated.}
}
\value{
A concatenated vector of 0s and 1s, where the length of each section
of the vector corresponds to the respective value in \code{ns}.
}
\description{
This function generates random binary vectors (composed of 0s and 1s) for
each element in the input vector \code{ns}. Each 0 and 1 is equally likely, with
the length of each binary vector determined by the corresponding element in
\code{ns}.
}
\examples{
# Generate binary vectors of lengths 3, 5, and 2
rgen01(c(3, 5, 2))

# Generate binary vectors of lengths 10, 20, and 30
rgen01(c(10, 20, 30))
}
