% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_fitting.R
\name{singleGLMMFit}
\alias{singleGLMMFit}
\title{Fit a single GLMM model}
\usage{
singleGLMMFit(formula, data, family, refit = TRUE, timeout = 3)
}
\arguments{
\item{formula}{A formula describing the model.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A character string representing the family (e.g., "nb" for Negative Binomial).}

\item{refit}{Logical indicating whethert a refit should be attempted if model does not converge.
Model with refitted with a different optimizer}

\item{timeout}{An integer or float indicating timeout for refitting if non-convergence}
}
\value{
A list with model parameters.
}
\description{
This function fits a GLMM using the specified formula, family, and data,
and extracts the relevant model parameters in a way that mirrors data generation by family.
}
