% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_fitting.R
\name{vcov.glmmfit}
\alias{vcov.glmmfit}
\title{Variance-Covariance Matrix for GLMM Fit}
\usage{
vcov.glmmfit(object, ...)
}
\arguments{
\item{object}{A fitted GLMM object of class \code{glmmfit}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix containing the variance-covariance matrix of the estimated parameters.
The row and column names are set to the parameter names based on the family of the model.
}
\description{
This function computes the Asymptotic variance-covariance matrix of the
regression parameters, family specific parameters and variance component of the random effect
}
\examples{
\dontrun{
  fit <- glmmfit(data, family = "binomial")
  vcov_matrix <- vcov(fit)
}

}
\seealso{
\code{\link[stats]{vcov}} for the generic variance-covariance method.
}
