% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{vcov.vpcObj}
\alias{vcov.vpcObj}
\title{Calculate the Variance-Covariance Matrix for a vpcObj}
\usage{
\method{vcov}{vpcObj}(object, ...)
}
\arguments{
\item{object}{An2 object of class \code{vpcObj} containing model information and gradients.}

\item{...}{Additional arguments (not used in this method).}
}
\value{
The variance of VPC obtained by Delta Method.
}
\description{
This function computes the variance-covariance matrix for an object of class \code{vpcObj}.
}
