% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2mat}
\alias{vec2mat}
\title{Convert Vector to Symmetric Matrix}
\usage{
vec2mat(vec)
}
\arguments{
\item{vec}{A numeric vector containing the elements to be placed in the lower triangular part of the matrix. The length of \code{vec} should be compatible with forming a symmetric matrix.}
}
\value{
A symmetric numeric matrix with dimensions determined by the length of \code{vec}.
}
\description{
This function converts a vector into a symmetric matrix by filling the lower triangular part of the matrix with the elements of the vector and mirroring it to the upper triangular part. The size of the matrix is automatically determined based on the length of the vector.
}
\examples{
# Convert a vector to a symmetric 3x3 matrix
vec <- c(2,1,2)
vec2mat(vec)

# Example with a longer vector for a larger matrix
vec <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
vec2mat(vec)
}
