% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVpc.R
\name{vpc.ga}
\alias{vpc.ga}
\title{Variance Partition Coefficient for Gaussian Model}
\usage{
vpc.ga(Sigma, phi, x, ...)
}
\arguments{
\item{Sigma}{A covariance matrix for the random effects.}

\item{phi}{The overdispersion parameter.}

\item{x}{The covariate values.}

\item{...}{Optional additional arguments.}
}
\value{
The VPC for the Gaussian model.
}
\description{
This function calculates the Variance Partition Coefficient (VPC)
for a Gaussian model, adjusting for random effects and overdispersion.
}
\examples{
Sigma <- matrix(c(0.1, 0.05, 0.05, 0.2), nrow = 2)
phi <- 1.5
x <- 0.7
vpc.ga(Sigma, phi, x)
}
