% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVpc.R
\name{vpc.nb}
\alias{vpc.nb}
\title{Variance Partition Coefficient for Negative Binomial Model}
\usage{
vpc.nb(beta, Sigma, phi, x, ...)
}
\arguments{
\item{beta}{A vector of regression coefficients.}

\item{Sigma}{A covariance matrix for the random effects.}

\item{phi}{The overdispersion parameter (can be overridden by \code{theta}).}

\item{x}{The covariate values.}

\item{...}{Optional additional arguments, such as \code{theta} to override \code{phi}.}
}
\value{
The VPC for the negative binomial model.
}
\description{
This function calculates the Variance Partition Coefficient (VPC)
for a negative binomial model, given the parameters.
}
\examples{
beta <- c(0.5, 1.2)
Sigma <- matrix(c(0.1, 0.05, 0.05, 0.2), nrow = 2)
phi <- 1.5
x <- 0.7
vpc.nb(beta, Sigma, phi, x)
}
