% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVpc.R
\name{vpc.tw}
\alias{vpc.tw}
\title{Variance Partition Coefficient for Tweedie (Compound Poisson)}
\usage{
vpc.tw(beta, Sigma, phi, power, x, ...)
}
\arguments{
\item{beta}{A vector of regression coefficients.}

\item{Sigma}{A covariance matrix for the random effects.}

\item{phi}{The overdispersion parameter.}

\item{power}{The power parameter.}

\item{x}{The covariate values.}

\item{...}{Optional additional arguments.}
}
\value{
The VPC for the transformed Weibull model.
}
\description{
This function calculates the Variance Partition Coefficient (VPC)
for a transformed Weibull model with a power transformation.
}
\examples{
beta <- c(0.5, 1.2)
Sigma <- matrix(c(0.1, 0.05, 0.05, 0.2), nrow = 2)
phi <- 1.5
power <- 2
x <- 0.7
vpc.tw(beta, Sigma, phi, power, x)
}
