% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_calculation.R
\name{vpc_from_paramMat}
\alias{vpc_from_paramMat}
\title{Calculate Variance Partition Coefficient (VPC) for Different Families}
\usage{
vpc_from_paramMat(family, paramMat, x)
}
\arguments{
\item{family}{A character string specifying the family for the model.
Can be one of
\code{"negative_binomial"} or \code{"tweedie"}.}

\item{paramMat}{A matrix or data frame containing the parameters for each
row. Columns
for \code{b0}, \code{b1}, \code{sig11}, \code{sig12}, \code{sig22},
and family-specific
parameters (such as \code{theta} for Negative Binomial or \code{phi},
\code{power} for Tweedie)
are required.}

\item{x}{Covariates or other necessary inputs for calculating the VPC.}
}
\value{
A list where each element corresponds to the VPC results for each row
of the parameter matrix.
}
\description{
This function takes a family type, a parameter matrix, and a covariate
(or covariates),
and calculates the VPC for each row of the parameter matrix. The function
dynamically
handles family-specific parameters such as \code{phi} and \code{power} for Tweedie
and Negative Binomial families.
}
