% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_ci.R
\name{adjustedc_mixture_ci}
\alias{adjustedc_mixture_ci}
\title{Adjusted Confidence Interval Calculation for VPC}
\usage{
adjustedc_mixture_ci(
  vpcObj,
  vpc.value,
  forcePD = F,
  alpha = 0.05,
  n = 1000,
  thresh = 0.01,
  truncated = TRUE
)
}
\arguments{
\item{vpcObj}{An object containing the VPC model with variance-covariance information.}

\item{vpc.value}{Numeric. The VPC value for which the confidence interval is to be computed.}

\item{forcePD}{whether the nearest positive matrix should be used}

\item{alpha}{Numeric. The significance level for the confidence interval (default is 0.05).}

\item{n}{number of bootstrap samples}

\item{thresh}{Integer. Definition of the boundary}

\item{truncated}{currently not used}
}
\value{
the confidence interval
}
\description{
Adjusted Confidence Interval Calculation for VPC
}
