% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{batchGLMMData}
\alias{batchGLMMData}
\title{Generate GLMM Data for Multiple Features}
\usage{
batchGLMMData(
  beta,
  ns,
  Sigma = NULL,
  num = 1,
  X = NULL,
  family = "gaussian",
  link = "identity",
  ...
)
}
\arguments{
\item{beta}{A vector of fixed effect coefficients.}

\item{ns}{A vector of sample sizes for each group.}

\item{Sigma}{A covariance matrix for the random effects.}

\item{num}{The number of features (Samples) to generate.}

\item{X}{A matrix of predictors (design matrix) for the fixed effects.}

\item{family}{A string specifying the distribution family
(default is "gaussian").Supported families include "gaussian", "poisson",
"negative_binomial", "tweedie", "gamma", and "binomial".}

\item{link}{A string specifying the link function (default is "identity").}

\item{...}{Additional arguments to be passed to the family-specific data
generation functions.}
}
\value{
A matrix where rows represent the predictors (design matrix \code{X})
and generated features. The feature rows are named as "Feature 1",
"Feature 2", etc.
}
\description{
This function generates data for multiple features using generalized linear
mixed models (GLMMs).
It calls the \code{singleGLMMData} function for each feature and combines the
results into a matrix.
}
\examples{
X <- matrix(1:10, nrow = 10)
beta <- c(0.5, 0.2)
Sigma <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
ns <- c(5,3,2)
batchGLMMData(beta = beta, ns = ns, Sigma = Sigma, num = 5, X = X,
              family = "gaussian", link = "identity")

}
