% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_test.R
\name{vpc.test.glmmfit}
\alias{vpc.test.glmmfit}
\alias{vpc.test.Glmmfits}
\alias{vpc.test}
\title{Perform Variance Partition Coefficient (VPC) Test}
\usage{
\method{vpc.test}{glmmfit}(
  fitObj,
  null_formula,
  type = c("classical", "self", "zhang", "julius")
)

\method{vpc.test}{Glmmfits}(
  fitObj,
  null_formula,
  type = c("classical", "self", "zhang", "julius")
)

vpc.test(
  fitObj,
  null_formula,
  type = c("classical", "self", "zhang", "julius")
)
}
\arguments{
\item{fitObj}{The fitted model object of class \code{glmmfit} or a list of fitted model objects
of class \code{glmmfits} for which the VPC is to be calculated.}

\item{null_formula}{A formula to specify a null model for comparison.}

\item{type}{method to chalculate the chi_square distribution.}
}
\value{
A numeric value or a list representing the calculated VPC for the given model,
including additional details if applicable.
}
\description{
This function calculates the Variance Partition Coefficient (VPC) for a fitted model object.
The specific method used will depend on the class of the provided model.
}
