% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst.R
\name{vstransform}
\alias{vstransform}
\title{Variance Stabilizing Transformation}
\usage{
vstransform(counts, blind = FALSE, num_cores = NULL)
}
\arguments{
\item{counts}{A matrix of integer count data.}

\item{blind}{Logical, whether to ignore the experimental design in the transformation.}

\item{num_cores}{Number of workers}
}
\value{
A matrix of transformed values.
}
\description{
Variance Stabilizing Transformation
}
\examples{
ns <- rep(10, 10)
X <- rgen01(ns)
beta <- c(3, 5)
Sigma <- matrix(c(2, 1, 1, 2), nrow=2)
family <- "nb"
cov_values <- c(0, 1)
formula <- y ~ x + (1 | group)
datafrmMat <- batchGLMMData(num=5, X, beta=beta, Sigma=Sigma, ns=ns,
family = "negative_binomial", theta = 1.5)
ys <- datafrmMat[-1,]; X <- datafrmMat[1,]; group <- colnames(datafrmMat)
vstransform(ys, num_cores=4)
}
