% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vstransform}
\alias{vstransform}
\title{Variance Stabilizing Transformation}
\usage{
vstransform(counts, blind = FALSE, num_cores = NULL)
}
\arguments{
\item{counts}{A matrix of integer count data.}

\item{blind}{Logical, whether to ignore the experimental design in the transformation.}

\item{num_cores}{Number of workers}
}
\value{
A matrix of transformed values.
}
\description{
Variance Stabilizing Transformation
}
